* created a unity project * moved source files to Packages folder * fixed dependencies to point to git repositories * version bump * use guid option in asmdef
