+++ Assets/Scenes.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 7b0dcce6685baa347a6192db2ea0c137
-TextScriptImporter:
+guid: d08bbd603c914964190713bd95689eb3
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
