+++ Assets/Scenes/SampleScene.unity.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 05a8ad35ad74cae48ab8aa03ac4298b8
-TextScriptImporter:
+guid: 9fc0d4010bbf28b4594072e72b8655ab
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
