+++ Packages/bindingrx/Examples/Scripts/BindingRx.Examples.asmdef.meta
