+++ Packages/bindingrx/Examples/Scripts/SliderExample.cs
