+++ Packages/bindingrx/Examples/Scripts/SliderExampleInstaller.cs
