+++ Packages/bindingrx/package.json
@@ -1,12 +1,12 @@
 {
   "name": "bindingrx",
   "displayName": "BindingRx",
-  "version": "3.0.0",
+  "version": "3.0.1",
   "unity": "2018.3",
   "description": "Data binding package for unity3d",
   "dependencies": {
     "type-inspector": "https://github.com/k0dep/type-inspector.git#1.1.0",
-    "unirx": "https://github.com/starikcetin/UniRx.git#7.1.0"
+    "com.neuecc.unirx": "https://github.com/starikcetin/UniRx.git#7.1.0"
   },
   "keywords": [
     "reflection",
