+++ Packages/bindingrx/package.json.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: c80844e5a660f5042ac653050d4482a3
-TextScriptImporter:
+PackageManifestImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
