+++ Packages/bindingrx/src/BindingWayType.cs
