+++ Packages/bindingrx/src/PropertyBinding.cs.meta
