+++ Packages/bindingrx/src/StateWatcher.cs
