+++ ProjectSettings/ProjectVersion.txt
@@ -1,0 +1,2 @@
+m_EditorVersion: 2019.2.10f1
+m_EditorVersionWithRevision: 2019.2.10f1 (923acd2d43aa)
